% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did_multiplegt_old.R
\name{did_multiplegt_old}
\alias{did_multiplegt_old}
\title{Old DID Multiplegt Estimator (de Chaisemartin and D'Haultfoeuille 2020)}
\usage{
did_multiplegt_old(
  df,
  Y,
  G,
  T,
  D,
  controls = c(),
  placebo = 0,
  dynamic = 0,
  threshold_stable_treatment = 0,
  recat_treatment = NULL,
  trends_nonparam = NULL,
  trends_lin = NULL,
  brep = 0,
  cluster = NULL,
  covariance = FALSE,
  average_effect = NULL,
  parallel = FALSE
)
}
\arguments{
\item{df}{the data frame for input}

\item{Y}{the name of Y variable}

\item{G}{the name of group variable}

\item{T}{the name of time variable}

\item{D}{the name of treatment variable}

\item{controls}{the list of names of control variables, empty if not specified}

\item{placebo}{the number of placebo estimators}

\item{dynamic}{the number of dynamic estimators}

\item{threshold_stable_treatment}{the threshold to mark a difference stable}

\item{recat_treatment}{the name of treatment group recategorized}

\item{trends_nonparam}{trends nonparametric option}

\item{trends_lin}{trends linear option}

\item{brep}{the number of bootstrap replications}

\item{cluster}{cluster variable for standard errors}

\item{covariance}{whether to compute covariance}

\item{average_effect}{average effect option}

\item{parallel}{whether to run bootstrap in parallel}
}
\value{
A list with treatment effects and related statistics
}
\description{
This is the old version of did_multiplegt from de Chaisemartin and D'Haultfoeuille (2020).
For most use cases, we recommend using did_multiplegt with mode="dyn" instead, which is faster and has more features.
}
\keyword{internal}
